chore(hlapi): move integrations tests into src dir
